# sql: rationalize batch key and bytes limit