--- a/pkg/sql/row/fetcher.go
+++ b/pkg/sql/row/fetcher.go
@@ -563,14 +563,51 @@ func (rf *Fetcher) GetTables() []catalog.Descriptor {
 	return ret
 }
 
+// RowLimit represents a response limit expressed in terms of number of result
+// rows. RowLimits get ultimately converted to KeyLimits and are translated into
+// BatchRequest.MaxSpanRequestKeys.
+type RowLimit uint64
+
+// KeyLimit represents a response limit expressed in terms of number of keys.
+type KeyLimit int64
+
+// BytesLimit represents a response limit expressed in terms of the size of the
+// results. A BytesLimit ultimately translates into BatchRequest.TargetBytes.
+type BytesLimit uint64
+
+// NoRowLimit can be passed to Fetcher.StartScan to signify that the caller
+// doesn't want to limit the number of result rows for each scan request.
+const NoRowLimit RowLimit = 0
+
+// NoBytesLimit can be passed to Fetcher.StartScan to signify that the caller
+// doesn't want to limit the size of results for each scan request.
+//
+// See also DefaultBatchBytesLimit.
+const NoBytesLimit BytesLimit = 0
+
 // StartScan initializes and starts the key-value scan. Can be used multiple
 // times.
+//
+// batchBytesLimit controls whether bytes limits are placed on the batches. If
+// set, bytes limits will be used to protect against running out of memory (on
+// both this client node, and on the server).
+//
+// If batchBytesLimit is set, rowLimitHint can also be set to control the number of
+// rows that will be scanned by the first batch. If set, subsequent batches (if
+// any) will have progressively higher limits (up to a fixed max). The idea with
+// row limits is to make the execution of LIMIT queries efficient: if the caller
+// has some idea about how many rows need to be read to ultimately satisfy the
+// query, the Fetcher uses it. Even if this hint proves insufficient, the
+// Fetcher continues to set row limits (in addition to bytes limits) on the
+// argument that some number of rows will eventually satisfy the query and we
+// likely don't need to scan `spans` fully. The bytes limit, on the other hand,
+// is simply intended to protect against OOMs.
 func (rf *Fetcher) StartScan(
 	ctx context.Context,
 	txn *kv.Txn,
 	spans roachpb.Spans,
-	limitBatches bool,
-	limitHint int64,
+	batchBytesLimit BytesLimit,
+	rowLimitHint RowLimit,
 	traceKV bool,
 	forceProductionKVBatchSize bool,
 ) error {
@@ -583,8 +620,8 @@ func (rf *Fetcher) StartScan(
 		makeKVBatchFetcherDefaultSendFunc(txn),
 		spans,
 		rf.reverse,
-		limitBatches,
-		rf.rowLimitToKeyLimit(limitHint),
+		batchBytesLimit,
+		rf.rowLimitToKeyLimit(rowLimitHint),
 		rf.lockStrength,
 		rf.lockWaitPolicy,
 		rf.lockTimeout,
@@ -621,8 +658,8 @@ func (rf *Fetcher) StartInconsistentScan(
 	initialTimestamp hlc.Timestamp,
 	maxTimestampAge time.Duration,
 	spans roachpb.Spans,
-	limitBatches bool,
-	limitHint int64,
+	batchBytesLimit BytesLimit,
+	rowLimitHint RowLimit,
 	traceKV bool,
 	forceProductionKVBatchSize bool,
 ) error {
@@ -683,8 +720,8 @@ func (rf *Fetcher) StartInconsistentScan(
 		sendFunc(sendFn),
 		spans,
 		rf.reverse,
-		limitBatches,
-		rf.rowLimitToKeyLimit(limitHint),
+		batchBytesLimit,
+		rf.rowLimitToKeyLimit(rowLimitHint),
 		rf.lockStrength,
 		rf.lockWaitPolicy,
 		rf.lockTimeout,
@@ -699,19 +736,19 @@ func (rf *Fetcher) StartInconsistentScan(
 	return rf.StartScanFrom(ctx, &f)
 }
 
-func (rf *Fetcher) rowLimitToKeyLimit(keyLimitHint int64) int64 {
-	if keyLimitHint == 0 {
+func (rf *Fetcher) rowLimitToKeyLimit(rowLimitHint RowLimit) KeyLimit {
+	if rowLimitHint == 0 {
 		return 0
 	}
 	// If we have a limit hint, we limit the first batch size. Subsequent
 	// batches get larger to avoid making things too slow (e.g. in case we have
 	// a very restrictive filter and actually have to retrieve a lot of rows).
-	// The keyLimitHint is a row limit, but each row could be made up of more than
+	// The rowLimitHint is a row limit, but each row could be made up of more than
 	// one key. We take the maximum possible keys per row out of all the table
 	// rows we could potentially scan over.
 	//
 	// We add an extra key to make sure we form the last row.
-	return keyLimitHint*int64(rf.maxKeysPerRow) + 1
+	return KeyLimit(int64(rowLimitHint)*int64(rf.maxKeysPerRow) + 1)
 }
 
 // StartScanFrom initializes and starts a scan from the given kvBatchFetcher. Can be