… pkg/sql/conn_executor_test.go
… pkg/sql/create_stats_test.go
… pkg/sql/execinfrapb/processors_sql.pb.go
… pkg/sql/indexbackfiller_test.go
… pkg/sql/logictest/testdata/logic_test/select_index_span_ranges
… pkg/sql/row/fetcher_test.go
… pkg/sql/row/kv_fetcher.go
… pkg/sql/rowexec/joinreader.go
… pkg/sql/rowexec/joinreader_blackbox_test.go
… pkg/sql/rowexec/rowfetcher.go
… pkg/sql/rowexec/stats.go
… pkg/sql/rowexec/tablereader.go
… pkg/sql/rowexec/zigzagjoiner.go
… pkg/sql/scan_test.go
… pkg/sql/stats/BUILD.bazel
… pkg/sql/stats/create_stats_job_test.go
… pkg/sql/tablewriter_delete.go
… pkg/testutils/lint/lint_test.go
… pkg/util/tracing/test_utils.go