@yuzefovich 2021-11-04
Hm, the increase of `ProductionKVBatchSize` from 10k to 100k is responsible for some micro-benchmark regressions. Here is an example of reducing the value back to 10k (the benchmark in `/pkg/bench`):
```
name                                                            old time/op    new time/op    delta
ScanFilter/Cockroach/count1=25/count2=400/limit=1-24              3.95ms ± 3%    1.20ms ± 3%  -69.59%  (p=0.000 n=10+10)
ScanFilter/Cockroach/count1=25/count2=400/limit=10-24              671µs ± 3%     670µs ± 2%     ~     (p=0.971 n=10+10)
ScanFilter/Cockroach/count1=25/count2=400/limit=50-24             2.64ms ± 4%    2.62ms ± 4%     ~     (p=0.393 n=10+10)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=1-24     4.95ms ± 1%    2.15ms ± 7%  -56.61%  (p=0.000 n=9+10)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=10-24    1.11ms ± 6%    1.10ms ± 4%     ~     (p=0.842 n=10+9)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=50-24    3.95ms ± 3%    3.95ms ± 4%     ~     (p=1.000 n=10+10)

name                                                            old alloc/op   new alloc/op   delta
ScanFilter/Cockroach/count1=25/count2=400/limit=1-24               596kB ± 0%     101kB ± 8%  -82.99%  (p=0.000 n=8+10)
ScanFilter/Cockroach/count1=25/count2=400/limit=10-24             46.8kB ± 3%    48.1kB ± 9%     ~     (p=0.190 n=9+9)
ScanFilter/Cockroach/count1=25/count2=400/limit=50-24              325kB ± 1%     330kB ± 6%     ~     (p=0.730 n=9+9)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=1-24      673kB ± 1%     175kB ±15%  -74.05%  (p=0.000 n=8+8)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=10-24     103kB ±99%      70kB ± 6%     ~     (p=0.211 n=10+9)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=50-24    1.27MB ± 2%    1.27MB ± 1%     ~     (p=0.863 n=9+9)

name                                                            old allocs/op  new allocs/op  delta
ScanFilter/Cockroach/count1=25/count2=400/limit=1-24                 511 ± 2%       506 ± 4%     ~     (p=0.052 n=8+9)
ScanFilter/Cockroach/count1=25/count2=400/limit=10-24                448 ± 0%       448 ± 0%     ~     (p=0.979 n=9+8)
ScanFilter/Cockroach/count1=25/count2=400/limit=50-24                883 ± 0%       883 ± 0%     ~     (p=0.748 n=9+8)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=1-24      1.28k ± 2%     1.21k ± 0%   -5.25%  (p=0.000 n=9+8)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=10-24       732 ±10%       712 ± 1%     ~     (p=0.116 n=9+9)
ScanFilter/MultinodeCockroach/count1=25/count2=400/limit=50-24     1.32k ± 2%     1.32k ± 1%     ~     (p=0.488 n=9+9)
```