@andreimatei 2021-11-04
What confuses me is why the higher limits seem to perform better than `limit=1`.